Refactor: Modify type

+++ src/main/java/com/example/whereareyou/schedule/response/ResponseTodaySchedule.java
@@ -6,5 +6,5 @@ import lombok.Data;
 @Data
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class ResponseTodaySchedule {
-    private long todaySchedule;
+    private int todaySchedule;
 }
